% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_Analysis}
\alias{df_Analysis}
\title{Simulated Analysis Cohort Dataset}
\format{
A data frame with 10,000 rows and 7 variables, ordered by variable "Q1":
\describe{
\item{ID}{A variable that uniquely identifies each unit and (potentially) used to link units in \code{df_Biomarker} and/or \code{df_Replicate}.}
\item{Time}{Simulated event time for each unit.}
\item{U}{Observed event time that is subject to random right-censoring.}
\item{delta}{Observed event time indicator: \eqn{I(T \leq C)}.}
\item{X}{Additional covariate used to generate the event time.}
\item{Z1}{True exposure variable for each unit. Note that this variable would not be observed in practice.}
\item{Q1}{The observed (noisy) exposure variable for each unit.}
}
}
\source{
{R script to regenerate simulated data can be found at \url{https://github.com/drtjt/NonLinearRC}}
}
\usage{
df_Analysis
}
\description{
A simulated \emph{analysis cohort} dataset to illustrate the use of \code{NL_RC}.
}
\examples{
data(df_Analysis)

}
\keyword{datasets}
