% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_Biomarker}
\alias{df_Biomarker}
\title{Simulated Biomarker Dataset}
\format{
A data frame with 500 rows and 5 variables:
\describe{
\item{ID}{A variable that uniquely identifies each unit and (potentially) used to link units in \code{df_Analysis} and/or \code{df_Replicate}.}
\item{W1}{The observed biomarker for each unit.}
\item{Q1}{The observed (noisy) exposure variable for each unit.}
\item{Z}{True exposure variable for each unit. Note that this variable would not be observed in practice.}
\item{X}{Additional covariate used in the measurement error model.}
}
}
\source{
{R script to regenerate simulated data can be found at \url{https://github.com/drtjt/NonLinearRC}}
}
\usage{
df_Biomarker
}
\description{
A simulated \emph{biomarker} dataset to illustrate the use of \code{NL_RC}.
}
\examples{
data(df_Biomarker)

}
\keyword{datasets}
