% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_Replicate}
\alias{df_Replicate}
\title{Simulated Replicate Dataset}
\format{
A data frame with 100 rows and 3 variables (two replicated observations of the biomarker for each unit):
\describe{
\item{ID}{A variable that uniquely identifies each unit and (potentially) used to link units in \code{df_Analysis} and/or \code{df_Biomarker}.}
\item{W1}{The first measurement of the biomarker for each unit.}
\item{W2}{The second measurement of the biomarker for each unit.}
}
}
\source{
{R script to regenerate simulated data can be found at \url{https://github.com/drtjt/NonLinearRC}}
}
\usage{
df_Replicate
}
\description{
A simulated \emph{replicate} dataset (in wide format) to illustrate the use of \code{NL_RC}.
}
\examples{
data(df_Replicate)

}
\keyword{datasets}
